Move new docs homepage components around
