+++ docs/blog/2026-04-16-introducing-thunder.mdx
@@ -1,8 +1,8 @@
 ---
-slug: introducing-thunder
-title: "Introducing Thunder: Lorem Ipsum Dolor Sit Amet"
-authors: [thunder-team]
-tags: [thunder, identity, security, "2026"]
+slug: introducing-thunderid
+title: "Introducing ThunderID: Lorem Ipsum Dolor Sit Amet"
+authors: [thunderid-team]
+tags: [thunderid, identity, security, "2026"]
 ---
 
 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
