+++ docs/blog/authors.yml
@@ -1,6 +1,6 @@
-thunder-team:
-  name: Thunder Team
-  title: Thunder Maintainers @ WSO2
+thunderid-team:
+  name: ThunderID Team
+  title: ThunderID Maintainers @ WSO2
   url: https://github.com/asgardeo/thunder
   image_url: https://github.com/asgardeo.png
   page: true
