+++ docs/content/guides/getting-started/configuration.mdx
@@ -411,7 +411,7 @@ Authorization resource settings.
 
 ### System Resource Server
 
-The system resource server controls administrative permissions for Thunder Console operations such as managing users, groups, organization units, and user types. You can customize the handle and identifier to fit your deployment requirements.
+The system resource server controls administrative permissions for <ProductName /> Console operations such as managing users, groups, organization units, and user types. You can customize the handle and identifier to fit your deployment requirements.
 
 | Setting | Default | Description |
 |---------|---------|-------------|
@@ -442,7 +442,7 @@ With the handle set to `mgmt`, all system permissions change as follows:
 
 #### Update Console Scopes
 
-When you change the system resource server handle, you must also update the Thunder Console scopes to match the new permission strings. The scopes tell the Console which permissions to request during sign-in.
+When you change the system resource server handle, you must also update the <ProductName /> Console scopes to match the new permission strings. The scopes tell the Console which permissions to request during sign-in.
 
 Update the `scopes` array in the Console configuration file for your deployment mode:
