+++ docs/package.json
@@ -21,15 +21,15 @@
     "wso2"
   ],
   "scripts": {
-    "build": "pnpm prebuild && docusaurus build",
+    "build": "pnpm setup && docusaurus build",
     "clear": "docusaurus clear",
     "deploy": "docusaurus deploy",
-    "dev": "pnpm prebuild && docusaurus start --port 5180",
+    "dev": "pnpm setup && docusaurus start --port 5180",
     "docusaurus": "docusaurus",
     "generate:api-specs": "node ./scripts/merge-openapi-specs.mjs",
     "generate:changelog": "node ./scripts/generate-changelog.mjs",
     "generate:postman-collections": "node ./scripts/generate-postman-collections.mjs",
-    "prebuild": "node ./scripts/prebuild.mjs",
+    "setup": "node ./scripts/prebuild.mjs",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "serve": "docusaurus serve",
@@ -46,22 +46,22 @@
     "@thunderid/eslint-plugin": "workspace:^",
     "@thunderid/logger": "workspace:^",
     "@thunderid/prettier-config": "workspace:^",
-    "@wso2/oxygen-ui": "catalog:",
-    "@wso2/oxygen-ui-icons-react": "catalog:",
+    "@wso2/oxygen-ui": "0.10.1",
+    "@wso2/oxygen-ui-icons-react": "0.10.1",
     "@scalar/api-reference-react": "0.9.15",
-    "clsx": "catalog:",
-    "eslint": "catalog:",
-    "prettier": "catalog:",
+    "clsx": "2.1.1",
+    "eslint": "9.39.4",
+    "prettier": "3.6.2",
     "prism-react-renderer": "2.3.0",
-    "react": "catalog:",
-    "react-dom": "catalog:"
+    "react": "19.2.3",
+    "react-dom": "19.2.3"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "3.9.2",
     "@docusaurus/tsconfig": "3.9.2",
     "@docusaurus/types": "3.9.2",
     "openapi-to-postmanv2": "4.23.0",
-    "typescript": "catalog:",
+    "typescript": "5.9.3",
     "yaml": "2.8.2"
   },
   "browserslist": {
